Move quat to euler decomposition to float3.
