+++ src/Tests/Tests/Shared/TestFloat3.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 50970075125c446eeaa24de1ba9a93b3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
